refactor: classes names changed, code restructure
