+++ restaurant/src/main/java/com/restaurantmanagement/controller/UserController.java
@@ -2,6 +2,7 @@ package com.restaurantmanagement.controller;
 
 
 
+import com.restaurantmanagement.service.impl.UserServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -13,7 +14,6 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import com.restaurantmanagement.security.model.User;
-import com.restaurantmanagement.service.UserService;
 
 import jakarta.validation.Valid;
 
@@ -26,10 +26,10 @@ public class UserController {
 
     private final Logger logger = LoggerFactory.getLogger(UserController.class);
 
-    private UserService userService;
+    private UserServiceImpl userService;
 
     @Autowired
-    public UserController(UserService userService) {
+    public UserController(UserServiceImpl userService) {
         this.userService = userService;
     }
     public int calculateFactorial(int number) {
