+++ restaurant/src/main/java/com/restaurantmanagement/security/model/User.java
@@ -19,6 +19,7 @@ public class User {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
+
   private String firstname;
   
   private String lastname;
