+++ restaurant/src/main/java/com/restaurantmanagement/service/UserServicePre.java
@@ -6,7 +6,7 @@ package com.restaurantmanagement.service;
 //import org.springframework.data.domain.Page;
 //import org.springframework.data.domain.Pageable;
 //
-//public interface UserService {
+//public interface UserServiceImpl {
 //
 //
 //    Page<User> getAllUsers(Pageable page);
