+++ restaurant/src/main/resources/application.properties
@@ -2,7 +2,7 @@
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/restaurantdb?useSSL=false&useUnicode=yes&characterEncoding=UTF-8&allowPublicKeyRetrieval=true&serverTimezone=UTC
 spring.datasource.username=root
-spring.datasource.password=
+spring.datasource.password=MyPassword123
 
 server.port=8090
